backend/requirements.txt
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0
